Create `Tabular` and move remaining domain-specific functionality
