+++ src/FastAI.jl
@@ -15,10 +15,8 @@ import DataAugmentation: getbounds, Bounds
 import DLPipelines: methoddataset, methodmodel, methodlossfn, methoddataloaders,
     mockmodel, mocksample, predict, predictbatch, mockmodel, encode, encodeinput,
     encodetarget, decodeŷ, decodey
-using IndirectArrays: IndirectArray
 using LearnBase: getobs, nobs
 using FilePathsBase
-using FixedPointNumbers
 using Flux
 using Flux.Optimise
 import Flux.Optimise: apply!, Optimiser, WeightDecay
@@ -38,9 +36,8 @@ import UnicodePlots
 using Statistics
 using InlineTest
 
-include("learner.jl")
 
-# Data block API
+# ## Data block API
 include("datablock/block.jl")
 include("datablock/encoding.jl")
 include("datablock/method.jl")
@@ -48,18 +45,19 @@ include("datablock/describe.jl")
 include("datablock/wrappers.jl")
 
 
-# submodules
-include("datasets/Datasets.jl")
-@reexport using .Datasets
+# ## Blocks
+# ### Wrapper blocks
+include("blocks/many.jl")
 
-include("models/Models.jl")
-using .Models
-
-# Blocks
+# ### Other
+include("blocks/continuous.jl")
 include("blocks/label.jl")
 
-# Encodings
-include("encodings/tabularpreprocessing.jl")
+# ## Encodings
+# ### Wrapper encodings
+include("encodings/only.jl")
+
+# ### Other
 include("encodings/onehot.jl")
 
 
@@ -67,24 +65,18 @@ include("encodings/onehot.jl")
 include("datablock/models.jl")
 include("datablock/loss.jl")
 
+
 # Interpretation
 include("interpretation/backend.jl")
 include("interpretation/text.jl")
-include("interpretation/detect.jl")
 include("interpretation/method.jl")
 include("interpretation/showinterpretable.jl")
 include("interpretation/learner.jl")
+include("interpretation/detect.jl")
 
 
-# Domain-specific
-include("Vision/Vision.jl")
-using .Vision
-
-include("Tabular/Tabular.jl")
-using .Tabular
-
-
-# training
+# Training
+include("learner.jl")
 include("training/paramgroups.jl")
 include("training/discriminativelrs.jl")
 include("training/utils.jl")
@@ -96,20 +88,41 @@ include("training/metrics.jl")
 include("serialization.jl")
 
 
+
+# submodules
+include("datasets/Datasets.jl")
+@reexport using .Datasets
+
+
 include("fasterai/methodregistry.jl")
 include("fasterai/learningmethods.jl")
 include("fasterai/defaults.jl")
 
 
+
+# Domain-specific
+include("Vision/Vision.jl")
+@reexport using .Vision
+export Image
+export Vision
+
+include("Tabular/Tabular.jl")
+@reexport using .Tabular
+
+
+include("interpretation/makie/stub.jl")
 function __init__()
     @require Makie="ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a" begin
         using .Makie
-        include("interpretation/makie/recipes.jl")
         include("interpretation/makie/showmakie.jl")
         include("interpretation/makie/lrfind.jl")
     end
 end
 
+module Models
+    using ..FastAI.Tabular: TabularModel
+    using ..FastAI.Vision.Models: xresnet18, xresnet50, UNetDynamic
+end
 
 
 export
@@ -135,11 +148,9 @@ export
     predictbatch,
 
     # blocks
-    Image,
-    Mask,
+
     Label,
     LabelMulti,
-    Keypoints,
     Many,
     TableRow,
     Continuous,
@@ -148,10 +159,7 @@ export
     encode,
     decode,
     setup,
-    ProjectiveTransforms,
-    ImagePreprocessing,
     OneHot,
-    KeypointPreprocessing,
     Only,
     Named,
     augs_projection, augs_lighting,
@@ -181,10 +189,6 @@ export
 
     # learning methods
     findlearningmethods,
-    ImageClassificationSingle,
-    ImageClassificationMulti,
-    ImageSegmentation,
-    ImageKeypointRegression,
     TabularClassificationSingle,
     TabularRegression,
