+++ src/Tabular/makie.jl
@@ -1,0 +1,1 @@
+# No Makie recipes yet, text is better I guess
