+++ src/Vision/learningmethods.jl
@@ -1,0 +1,191 @@
+
+function ImageClassificationSingle(
+		blocks::Tuple{<:Image{N},<:Label},
+        data=nothing;
+		size=ntuple(i -> 128, N),
+		aug_projections=DataAugmentation.Identity(),
+		aug_image=DataAugmentation.Identity(),
+        C=RGB{N0f8},
+        computestats=false,
+	) where N
+	return BlockMethod(
+		blocks,
+		(
+			ProjectiveTransforms(size; augmentations=aug_projections),
+			getimagepreprocessing(data, computestats; C=C, augmentations=aug_image),
+        	OneHot()
+		)
+	)
+end
+
+
+"""
+    ImageClassificationSingle(size, classes; kwargs...)
+    ImageClassificationSingle(blocks[, data]; kwargs...)
+
+Learning method for single-label image classification. Images are
+resized to `size` and classified into one of `classes`.
+
+Use [`ImageClassificationMulti`](#) for the multi-class setting.
+
+## Keyword arguments
+
+- `computestats = false`: Whether to compute image statistics on dataset `data` or use
+    default ImageNet stats.
+- `aug_projections = `[`DataAugmentation.Identity`](#): augmentation to apply during
+  [`ProjectiveTransforms`](#) (resizing and cropping)
+- `aug_image = `[`DataAugmentation.Identity`](#): pixel-level augmentation to apply during
+  [`ImagePreprocessing`](#)
+- `C = RGB{N0f8}`: Color type images are converted to before further processing. Use `Gray{N0f8}`
+    for grayscale images.
+"""
+function ImageClassificationSingle(size::NTuple{N,Int}, classes::AbstractVector; kwargs...) where N
+    blocks = (Image{N}(), Label(classes))
+    return ImageClassificationSingle(blocks, size=size)
+end
+
+registerlearningmethod!(FASTAI_METHOD_REGISTRY, ImageClassificationSingle, (Image, Label))
+
+# ---
+
+function ImageClassificationMulti(
+		blocks::Tuple{<:Image{N},<:LabelMulti},
+        data = nothing;
+		size=ntuple(i -> 128, N),
+		aug_projections=DataAugmentation.Identity(),
+		aug_image=DataAugmentation.Identity(),
+        C=RGB{N0f8},
+        computestats=false,
+	) where N
+	return BlockMethod(
+		blocks,
+		(
+			ProjectiveTransforms(size; augmentations=aug_projections),
+			getimagepreprocessing(data, computestats; C=C, augmentations=aug_image),
+        	OneHot()
+		)
+	)
+end
+
+
+"""
+    ImageClassificationMulti(size, classes; kwargs...)
+
+Learning method for multi-label image classification. Images are
+resized to `size` and classified into multiple of `classes`.
+
+Use [`ImageClassificationSingle`](#) for the single-class setting.
+
+## Keyword arguments
+
+- `computestats = false`: Whether to compute image statistics on dataset `data` or use
+    default ImageNet stats.
+- `aug_projections = `[`DataAugmentation.Identity`](#): augmentation to apply during
+  [`ProjectiveTransforms`](#) (resizing and cropping)
+- `aug_image = `[`DataAugmentation.Identity`](#): pixel-level augmentation to apply during
+  [`ImagePreprocessing`](#)
+- `C = RGB{N0f8}`: Color type images are converted to before further processing. Use `Gray{N0f8}`
+    for grayscale images.
+"""
+function ImageClassificationMulti(size::NTuple{N,Int}, classes::AbstractVector; kwargs...) where N
+    blocks = (Image{N}(), LabelMulti(classes))
+    return ImageClassificationMulti(blocks; size=size, kwargs...)
+end
+
+
+registerlearningmethod!(FASTAI_METHOD_REGISTRY, ImageClassificationMulti, (Image, LabelMulti))
+
+# ---
+
+function ImageSegmentation(
+		blocks::Tuple{<:Image{N},<:Mask{N}},
+        data=nothing;
+		size=ntuple(i -> 128, N),
+		aug_projections=DataAugmentation.Identity(),
+		aug_image=DataAugmentation.Identity(),
+        C=RGB{N0f8},
+        computestats=false,
+	) where N
+	return BlockMethod(
+		blocks,
+		(
+			ProjectiveTransforms(size; augmentations=aug_projections),
+			getimagepreprocessing(data, computestats; C=C, augmentations=aug_image),
+        	OneHot()
+		)
+	)
+end
+
+
+"""
+    ImageSegmentation(size, classes; kwargs...)
+
+Learning method for image segmentation. Images are
+resized to `size` and a class is predicted for every pixel.
+
+## Keyword arguments
+
+- `computestats = false`: Whether to compute image statistics on dataset `data` or use
+    default ImageNet stats.
+- `aug_projections = `[`DataAugmentation.Identity`](#): augmentation to apply during
+  [`ProjectiveTransforms`](#) (resizing and cropping)
+- `aug_image = `[`DataAugmentation.Identity`](#): pixel-level augmentation to apply during
+  [`ImagePreprocessing`](#)
+- `C = RGB{N0f8}`: Color type images are converted to before further processing. Use `Gray{N0f8}`
+    for grayscale images.
+"""
+function ImageSegmentation(size::NTuple{N,Int}, classes::AbstractVector; kwargs...) where N
+    blocks = (Image{N}(), Mask{N}(classes))
+    return ImageSegmentation(blocks; size=size, kwargs...)
+end
+
+registerlearningmethod!(FASTAI_METHOD_REGISTRY, ImageSegmentation, (Image, Mask))
+
+
+# ---
+
+
+function ImageKeypointRegression(
+        blocks::Tuple{<:Image{N},<:Keypoints{N}},
+        data=nothing;
+        size=ntuple(i -> 128, N),
+		aug_projections=DataAugmentation.Identity(),
+		aug_image=DataAugmentation.Identity(),
+        C=RGB{N0f8},
+        computestats=false,
+    ) where N
+    return BlockMethod(
+        blocks,
+        (
+			ProjectiveTransforms(size; augmentations=aug_projections),
+			getimagepreprocessing(data, computestats; C=C, augmentations=aug_image),
+            KeypointPreprocessing(size),
+        )
+    )
+end
+
+
+"""
+    ImageKeypointRegression(size, nkeypoints; kwargs...)
+
+Learning method for regressing a set of `nkeypoints` keypoints from
+images. Images are resized to `size` and a class is predicted for every pixel.
+"""
+function ImageKeypointRegression(size::NTuple{N,Int}, nkeypoints::Int; kwargs...) where N
+    blocks = (Image{N}(), Keypoints{N}((nkeypoints,)))
+    return ImageKeypointRegression(blocks; size = size, kwargs...)
+end
+
+registerlearningmethod!(FASTAI_METHOD_REGISTRY, ImageKeypointRegression, (Image, Keypoints))
+
+
+function getimagepreprocessing(data, computestats::Bool; kwargs...)
+    if isnothing(data) && computestats
+        error("If `computestats` is `true`, you have to pass in a data container `data`.")
+    end
+    return if computestats
+        setup(ImagePreprocessing, Image{2}(), data; kwargs...)
+    else
+        ImagePreprocessing(; kwargs...)
+    end
+end
