+++ src/blocks/continuous.jl
@@ -1,0 +1,22 @@
+
+"""
+    Continuous(size) <: Block
+
+`Block` for collections of numbers. `data` is valid if it's
+length is `size` and contains `Number`s.
+"""
+struct Continuous <: Block
+    size::Int
+end
+
+function checkblock(block::Continuous, x)
+    block.size == length(x) && eltype(x) <: Number
+end
+
+mockblock(block::Continuous) = rand(block.size)
+
+
+function blocklossfn(outblock::Continuous, yblock::Continuous)
+    outblock.size == yblock.size || error("Sizes of $outblock and $yblock differ!")
+    return Flux.Losses.mse
+end
