+++ src/interpretation/makie/showmakie.jl
@@ -2,24 +2,9 @@
 using FastAI
 import FastAI: createhandle, showblock!
 
-"""
-    ShowMakie([; kwargs...]) <: ShowBackend
-
-A backend for showing block data that uses
-[Makie.jl](https://github.com/JuliaPlots/Makie.jl) figures for
-visualization.
-
-Keyword arguments are passed through to the constructed `Figure`s.
-"""
-struct ShowMakie <: ShowBackend
-    size::Tuple{Int,Int}
-    kwargs
-end
-ShowMakie(sz=(500, 500); kwargs...) = ShowMakie(sz, kwargs)
-
 
 function createhandle(backend::ShowMakie; kwargs...)
-    fig = Figure(;kwargs..., backend.kwargs...)
+    fig = Figure(; kwargs..., backend.kwargs...)
     grid = fig[1, 1] = GridLayout()
     return grid
 end
@@ -31,7 +16,7 @@ function showblock(backend::ShowMakie, block, data)
     height = round(Int, 1.1 * backend.size[1])
 
 
-    grid = createhandle(backend, resolution=(width, height))
+    grid = createhandle(backend, resolution = (width, height))
     fig = grid.parent.parent
     showblock!(grid, backend, block, data)
     return fig
@@ -58,18 +43,18 @@ function showblock!(grid, backend::ShowMakie, blocks::Tuple, datas::Tuple)
     col = 1
     for (i, (block, data)) in enumerate(zip(blocks, datas))
         w = _nblocks(block)
-        subgrid = grid[1, col:col+w-1] = GridLayout(tellheight=false)
+        subgrid = grid[1, col:col+w-1] = GridLayout(tellheight = false)
         showblock!(subgrid, backend, block, data)
-        for j in col:col+w-1
+        for j = col:col+w-1
             colsize!(grid, j, Makie.Fixed(backend.size[1]))
         end
         col += w
     end
 
     # Add titles to named blocks
-    Makie.Label(grid[0, 1], "", tellwidth=false, textsize=25)
+    Makie.Label(grid[0, 1], "", tellwidth = false, textsize = 25)
     for (i, title) in enumerate(header)
-        Makie.Label(grid[1, i], title, tellwidth=false, textsize=25)
+        Makie.Label(grid[1, i], title, tellwidth = false, textsize = 25)
     end
 
 end
@@ -78,9 +63,9 @@ end
 function showblocks(backend::ShowMakie, block, datas)
     width = _nblocks(block) * backend.size[2]
     height = round(Int, length(datas) * 1.1 * backend.size[1])
-    res = (width*1.2, height*1.1)
+    res = (width * 1.2, height * 1.1)
 
-    grid = createhandle(backend, resolution=res)
+    grid = createhandle(backend, resolution = res)
     fig = grid.parent.parent
 
     showblocks!(grid, backend, block, datas)
@@ -96,21 +81,21 @@ function showblocks!(grid, backend::ShowMakie, blocks::Tuple, datas::AbstractVec
 
     # Show each sample in one row
     for (i, data) in enumerate(datas)
-        subgrid = grid[i, 1:n] = GridLayout(tellheight=false)
+        subgrid = grid[i, 1:n] = GridLayout(tellheight = false)
         rowsize!(grid, i, Makie.Fixed(backend.size[2]))
         showblock!(subgrid, backend, blocks, data)
     end
 
-    for i in 1:n
+    for i = 1:n
         colsize!(grid, i, Makie.Fixed(backend.size[1]))
     end
 
     # Add titles to named blocks
-    Makie.Label(grid[0, 1], "", textsize=25)
+    Makie.Label(grid[0, 1], "", textsize = 25)
     col = 1
     for (i, (title, block)) in enumerate(zip(header, blocks))
         w = _nblocks(block)
-        Makie.Label(grid[1, col:col+w-1], title, tellwidth=false, textsize=25)
+        Makie.Label(grid[1, col:col+w-1], title, tellwidth = false, textsize = 25)
         col += w
     end
 end
@@ -121,54 +106,28 @@ showblocks!(grid, backend::ShowMakie, block, datas::AbstractVector) =
 
 
 function showblock!(grid, ::ShowMakie, block::Label, data)
-    ax = imageaxis(grid[1, 1])
-    text!(ax, string(data), space=:data)
+    ax = cleanaxis(grid[1, 1])
+    text!(ax, string(data), space = :data)
 end
 
 function showblock!(grid, ::ShowMakie, block::LabelMulti, data)
-    ax = imageaxis(grid[1, 1])
-    text!(ax, join(string.(data), "\n"), space=:data)
+    ax = cleanaxis(grid[1, 1])
+    text!(ax, join(string.(data), "\n"), space = :data)
 end
 
 
-function showblock!(grid, ::ShowMakie, block::Union{<:OneHotTensor{0},<:OneHotTensorMulti{0}}, data)
+function showblock!(
+    grid,
+    ::ShowMakie,
+    block::Union{<:OneHotTensor{0},<:OneHotTensorMulti{0}},
+    data,
+)
     if !(sum(data) ≈ 1)
         data = softmax(data)
     end
-    ax = Axis(grid[1, 1], yticks=(1:length(block.classes), string.(block.classes),))
-    barplot!(
-		ax,
-		data,
-		direction=:x,
-	)
+    ax = Axis(grid[1, 1], yticks = (1:length(block.classes), string.(block.classes)))
+    barplot!(ax, data, direction = :x)
     hidespines!(ax)
-end
-
-
-function showblock!(grid, ::ShowMakie, block::Image{2}, data)
-    ax = imageaxis(grid[1, 1])
-    plotimage!(ax, data)
-end
-function showblock!(grid, ::ShowMakie, block::Mask{2}, data)
-    ax = imageaxis(grid[1, 1])
-    plotmask!(ax, data, block.classes)
-end
-
-
-function showblock!(grid, ::ShowMakie, block::Keypoints{2}, data)
-    ax = imageaxis(grid[1, 1])
-    h = maximum(first.(data))
-    ks = [SVector(x, h-y) for (y, x) in data]
-    scatter!(ax, ks)
-end
-
-function showblock!(grid, ::ShowMakie, block::Bounded{2, <:Keypoints{2}}, data)
-    ax = imageaxis(grid[1, 1])
-    h, w = block.size
-    ks = [SVector(x, h-y) for (y, x) in data]
-    xlims!(ax, 0, w)
-    ylims!(ax, 0, h)
-    scatter!(ax, ks)
 end
 
 
@@ -178,4 +137,31 @@ function default_showbackend()
     else
         return ShowMakie()
     end
+end
+
+
+## Helpers
+
+"""
+    cleanaxis(f)
+
+Create a `Makie.Axis` with no interactivity, decorations and aspect distortion.
+"""
+function cleanaxis(f; kwargs...)
+    ax = Makie.Axis(f; kwargs...)
+    ax.aspect = Makie.DataAspect()
+    ax.xzoomlock = true
+    ax.yzoomlock = true
+    ax.xrectzoom = false
+    ax.yrectzoom = false
+    ax.panbutton = nothing
+    ax.xpanlock = true
+    ax.ypanlock = true
+    ax.bottomspinevisible = false
+    ax.leftspinevisible = false
+    ax.rightspinevisible = false
+    ax.topspinevisible = false
+    Makie.hidedecorations!(ax)
+
+    return ax
 end
