--- test/encodings/many.jl
@@ -1,8 +1,0 @@
-include("../imports.jl")
-
-
-@testset ExtendedTestSet "`Many`" begin
-    enc = ImagePreprocessing()
-    block = Many(Image{2}())
-    FastAI.testencoding(enc, block)
-end
