+++ test/imports.jl
@@ -3,8 +3,9 @@ using Colors: RGB, N0f8, Gray
 using FastAI
 using FastAI: ParamGroups, IndexGrouper, getgroup, DiscriminativeLRs, decay_optim
 import FastAI: Image, Keypoints, Mask, testencoding, Label, OneHot, ProjectiveTransforms,
-    encodedblock, decodedblock, encode, decode, mockblock, checkblock, Block, Encoding,
-    imagedatasetstats
+    encodedblock, decodedblock, encode, decode, mockblock, checkblock, Block, Encoding
+import FastAI.Tabular: EncodedTableRow
+using FastAI.Vision: imagedatasetstats
 import Makie
 using FilePathsBase
 using FastAI.Datasets
