+++ test/runtests.jl
@@ -16,18 +16,6 @@ include("imports.jl")
     end
 
     @testset ExtendedTestSet "encodings/" begin
-        @testset ExtendedTestSet "projective.jl" begin
-            include("encodings/projective.jl")
-        end
-        @testset ExtendedTestSet "imagepreprocessing.jl" begin
-            include("encodings/imagepreprocessing.jl")
-        end
-        @testset ExtendedTestSet "keypointpreprocessing.jl" begin
-            include("encodings/keypointpreprocessing.jl")
-        end
-        @testset ExtendedTestSet "many.jl" begin
-            include("encodings/many.jl")
-        end
         @testset ExtendedTestSet "tabularpreprocessing.jl" begin
             include("encodings/tabularpreprocessing.jl")
         end
